Fix generate squares dictionary
Update documentation
