+++ CODEOWNERS
@@ -1,1 +1,2 @@
 * @samthegitguy
+src/arithmetic.js src/primes.js src/squares.js @samthegitguy
