+++ docs/global.html
@@ -104,7 +104,7 @@
     
 
     
-    <h4 class="name" id="foo"><span class="type-signature"></span>foo<span class="signature">()</span><span class="type-signature"></span></h4>
+    <h4 class="name" id="generateSquaresDictionary"><span class="type-signature"></span>generateSquaresDictionary<span class="signature">(length)</span><span class="type-signature"></span></h4>
     
 
     
@@ -112,7 +112,7 @@
 
 
 <div class="description">
-    foo function
+    Generates a dictionary of squares, until the length provided.Returns dictionary square:rootTherefore
 </div>
 
 
@@ -121,6 +121,204 @@
 
 
 
+
+
+    <h5>Parameters:</h5>
+    
+
+<table class="params">
+    <thead>
+    <tr>
+        
+        <th>Name</th>
+        
+
+        <th>Type</th>
+
+        
+
+        
+
+        <th class="last">Description</th>
+    </tr>
+    </thead>
+
+    <tbody>
+    
+
+        <tr>
+            
+                <td class="name"><code>length</code></td>
+            
+
+            <td class="type">
+            
+            </td>
+
+            
+
+            
+
+            <td class="description last">Length of array. Recommended to keep it same as the number you are rooting.</td>
+        </tr>
+
+    
+    </tbody>
+</table>
+
+
+
+
+
+
+<dl class="details">
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+    <dt class="tag-author">Author:</dt>
+    <dd class="tag-author">
+        <ul>
+            <li>samthegitguy</li>
+        </ul>
+    </dd>
+    
+
+    
+
+    
+
+    
+
+    
+    <dt class="tag-source">Source:</dt>
+    <dd class="tag-source"><ul class="dummy"><li>
+        <a href="squares.js.html">squares.js</a>, <a href="squares.js.html#line16">line 16</a>
+    </li></ul></dd>
+    
+
+    
+
+    
+
+    
+</dl>
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+<h5>Returns:</h5>
+
+        
+<div class="param-desc">
+    Dictionary of squares
+</div>
+
+
+
+    
+
+
+
+
+    <h5>Example</h5>
+    
+    <pre class="prettyprint"><code>squares.generateSquaresDictionary(100)[100]// returns 10squares.generateSquaresDictionary(25)[25]// returns 5&lt;caption>Returns the square root of the number you put in&lt;/caption>NOTE: This does not work on non-square numbers - not every number is in the dict, only square numbers. Use squareroot() instead.</code></pre>
+
+
+
+        
+            
+
+    
+
+    
+    <h4 class="name" id="squareroot"><span class="type-signature"></span>squareroot<span class="signature">(num)</span><span class="type-signature"></span></h4>
+    
+
+    
+
+
+
+<div class="description">
+    Generates the square root of all positive numbers using generateSquares.
+</div>
+
+
+
+
+
+
+
+
+
+    <h5>Parameters:</h5>
+    
+
+<table class="params">
+    <thead>
+    <tr>
+        
+        <th>Name</th>
+        
+
+        <th>Type</th>
+
+        
+
+        
+
+        <th class="last">Description</th>
+    </tr>
+    </thead>
+
+    <tbody>
+    
+
+        <tr>
+            
+                <td class="name"><code>num</code></td>
+            
+
+            <td class="type">
+            
+            </td>
+
+            
+
+            
+
+            <td class="description last">The number which needs to be squarerooted.</td>
+        </tr>
+
+    
+    </tbody>
+</table>
 
 
 
@@ -156,7 +354,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="index.js.html">index.js</a>, <a href="index.js.html#line2">line 2</a>
+        <a href="squares.js.html">squares.js</a>, <a href="squares.js.html#line28">line 28</a>
     </li></ul></dd>
     
 
@@ -181,6 +379,16 @@
 
 
 
+<h5>Returns:</h5>
+
+        
+<div class="param-desc">
+    The square root of a number
+</div>
+
+
+
+    
 
 
 
@@ -202,13 +410,13 @@
 </div>
 
 <nav>
-    <h2><a href="index.html">Home</a></h2><h3>Global</h3><ul><li><a href="global.html#foo">foo</a></li></ul>
+    <h2><a href="index.html">Home</a></h2><h3>Global</h3><ul><li><a href="global.html#generateSquaresDictionary">generateSquaresDictionary</a></li><li><a href="global.html#squareroot">squareroot</a></li></ul>
 </nav>
 
 <br class="clear">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc/jsdoc">JSDoc 3.6.7</a> on Fri Jul 09 2021 18:35:14 GMT+1000 (Australian Eastern Standard Time)
+    Documentation generated by <a href="https://github.com/jsdoc/jsdoc">JSDoc 3.6.7</a> on Sun Jul 11 2021 11:29:04 GMT+1000 (Australian Eastern Standard Time)
 </footer>
 
 <script> prettyPrint(); </script>
