+++ docs/squares.js.html
@@ -1,0 +1,84 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="utf-8">
+    <title>JSDoc: Source: squares.js</title>
+
+    <script src="scripts/prettify/prettify.js"> </script>
+    <script src="scripts/prettify/lang-css.js"> </script>
+    <!--[if lt IE 9]>
+      <script src="//html5shiv.googlecode.com/svn/trunk/html5.js"></script>
+    <![endif]-->
+    <link type="text/css" rel="stylesheet" href="styles/prettify-tomorrow.css">
+    <link type="text/css" rel="stylesheet" href="styles/jsdoc-default.css">
+</head>
+
+<body>
+
+<div id="main">
+
+    <h1 class="page-title">Source: squares.js</h1>
+
+    
+
+
+
+    
+    <section>
+        <article>
+            <pre class="prettyprint source linenums"><code>const square = a => a * a
+
+/** Generates a dictionary of squares, until the length provided.
+ * Returns dictionary square:root
+ * Therefore 
+ * @example
+ * squares.generateSquaresDictionary(100)[100]
+ * // returns 10
+ * squares.generateSquaresDictionary(25)[25]
+ * // returns 5
+ * &lt;caption>Returns the square root of the number you put in&lt;/caption>
+ * NOTE: This does not work on non-square numbers - not every number is in the dict, only square numbers. Use squareroot() instead.
+ * @author samthegitguy
+ * @argument length Length of array. Recommended to keep it same as the number you are rooting.
+ * @returns Dictionary of squares **/
+const generateSquaresDictionary = function(length) {
+    let squares = {}
+    for(let i = 0; i &lt; length + 1; i++) {
+        squares[square(i)] = i
+    }
+    return squares
+}
+/**
+ * Generates the square root of all positive numbers using generateSquares.
+ * @param num The number which needs to be squarerooted.
+ * @returns The square root of a number
+ */
+function squareroot(num) {
+    return generateSquaresDictionary(num)[num]
+}
+// Exports
+exports.square = square
+exports.generateSquares = generateSquaresDictionary
+exports.squareroot = squareroot</code></pre>
+        </article>
+    </section>
+
+
+
+
+</div>
+
+<nav>
+    <h2><a href="index.html">Home</a></h2><h3>Global</h3><ul><li><a href="global.html#generateSquaresDictionary">generateSquaresDictionary</a></li><li><a href="global.html#squareroot">squareroot</a></li></ul>
+</nav>
+
+<br class="clear">
+
+<footer>
+    Documentation generated by <a href="https://github.com/jsdoc/jsdoc">JSDoc 3.6.7</a> on Sun Jul 11 2021 11:29:04 GMT+1000 (Australian Eastern Standard Time)
+</footer>
+
+<script> prettyPrint(); </script>
+<script src="scripts/linenumber.js"> </script>
+</body>
+</html>
