+++ src/arithmetic.js
