+++ src/primes.js
@@ -1,0 +1,3 @@
+function findPrime(num) {
+    
+}
