+++ tests/arithmetic.test.js
@@ -1,5 +1,5 @@
 // Packages
-const arithmetic = require("../src/arithmetic.cjs")
+const arithmetic = require("../src/arithmetic.js")
 
 test('adds 1 + 2 to equal 3', () => expect(arithmetic.add(1,2)).toEqual(3))
 test('subtracts 1 from 2 to equal 1', () => expect(arithmetic.subtract(2,1)).toEqual(1))
